src/portfolio-reader-svc/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
